refactor matchtype
